Fixed duplicated code about convertion of void* to hex with function such as ToHex(). Fixed instance address printing in LoggerHelper class. Updated code that print a Menu description.
